backend/requirements.txt
Django>=5.2,<6.0
django-cors-headers
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0
Pillow>=10.0.0
django-filter>=24.2
python-dateutil>=2.9.0
